+++ src/views/io_chart.h
@@ -14,7 +14,7 @@ struct IoChartData {
 struct IoChartState {
   BumpArena cur_arena;
   GrowingArray<IoChartData> charts;
-  size_t wasted_bytes;
+  uint32_t wasted_bytes;
 };
 
 void io_chart_update(IoChartState &my_state, const State &state);
