+++ src/views/library_viewer.h
@@ -40,7 +40,7 @@ struct LibraryViewerWindow {
 struct LibraryViewerState {
   GrowingArray<LibraryViewerWindow> windows;
   BumpArena cur_arena;
-  size_t updates_since_last_cleanup;
+  uint32_t updates_since_last_cleanup;
 };
 
 struct FrameContext;
